auth registration, confirmation, redis and postgres connection

+++ backend/Entities/Day.cs
@@ -6,7 +6,7 @@ namespace backend.Entities;
 public class Day
 {
     [Key][JsonIgnore] public int Id { get; set; }
-    public float DateMark { get; set; }
+    public double DateMark { get; set; }
     public User User { get; set; }
     public int dayNumber { get; set; }
     // public ICollection<Event> Events { get; set; }
